{"schema":"decodex/commit/1","summary":"Split service list modules","authority":"manual"}
